refactor: remove PrismaClientFactory
