+++ src/modules/prisma/prisma.types.ts
@@ -1,10 +1,34 @@
 import type { SingleKeyMap } from '@douglasneuroinformatics/libjs';
 import type { Prisma, PrismaClient } from '@prisma/client';
+import type {
+  Call,
+  DefaultArgs,
+  DynamicClientExtensionThis,
+  InternalArgs,
+  MergeExtArgs
+} from '@prisma/client/runtime/library';
 import type { IfNever } from 'type-fest';
 
 import type { UserConfig } from '../../user-config.js';
 import type { DefaultPrismaClientOptions } from './prisma.config.js';
-import type { ExtendedPrismaClient } from './prisma.factory.js';
+import type { ModelExtArgs, ResultExtArgs } from './prisma.extensions.js';
+
+type InferPrismaExtensionArgs<TArgs extends { [key: string]: unknown }> = MergeExtArgs<
+  Prisma.TypeMap<DefaultArgs, RuntimePrismaClientOptions['omit']>,
+  {},
+  InternalArgs<TArgs['result'], TArgs['model'], TArgs['query'], TArgs['client']>
+>;
+
+type InferExtendedClient<TArgs extends { [key: string]: unknown }> = DynamicClientExtensionThis<
+  Call<
+    Prisma.TypeMapCb<RuntimePrismaClientOptions>,
+    {
+      extArgs: InferPrismaExtensionArgs<TArgs>;
+    }
+  >,
+  Prisma.TypeMapCb<RuntimePrismaClientOptions>,
+  InferPrismaExtensionArgs<TArgs>
+>;
 
 export type PrismaClientLike = PrismaClient & {
   [key: string]: any;
@@ -30,3 +54,5 @@ export type RuntimePrismaClientOptions = UserConfig extends {
 }
   ? TPrismaClientOptions
   : DefaultPrismaClientOptions;
+
+export type ExtendedPrismaClient = InferExtendedClient<{ model: ModelExtArgs; result: ResultExtArgs }>;
